+++ packages/cli/src/commands/agent-runs/format.ts
