+++ packages/cli/src/util/telemetry/commands/agent-runs/projects.ts
@@ -1,5 +1,5 @@
 import type { TelemetryMethods } from '../../types';
-import type { projectsSubcommand } from '../../../../commands/agent/command';
+import type { projectsSubcommand } from '../../../../commands/agent-runs/command';
 import { AgentRunsQueryTelemetryClient } from './shared';
 
 export class AgentProjectsTelemetryClient
