+++ packages/cli/src/util/telemetry/commands/agent/index.ts
@@ -6,38 +6,18 @@ export class AgentTelemetryClient
   extends TelemetryClient
   implements TelemetryMethods<typeof agentCommand>
 {
-  trackCliSubcommandInit(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'init',
-      value: actual,
-    });
+  trackCliArgumentInit(value: string | undefined) {
+    if (value) {
+      this.trackCliArgument({
+        arg: 'init',
+        value,
+      });
+    }
   }
 
-  trackCliSubcommandRuns(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'runs',
-      value: actual,
-    });
-  }
-
-  trackCliSubcommandInspect(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'inspect',
-      value: actual,
-    });
-  }
-
-  trackCliSubcommandTrace(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'trace',
-      value: actual,
-    });
-  }
-
-  trackCliSubcommandProjects(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'projects',
-      value: actual,
-    });
+  trackCliFlagYes(value: boolean | undefined) {
+    if (value) {
+      this.trackCliFlag('yes');
+    }
   }
 }
