+++ packages/cli/src/util/telemetry/root.ts
@@ -19,6 +19,13 @@ export class RootTelemetryClient extends TelemetryClient {
     });
   }
 
+  trackCliCommandAgentRuns(actual: string) {
+    this.trackCliCommand({
+      command: 'agent-runs',
+      value: actual,
+    });
+  }
+
   trackCliCommandAiGateway(actual: string) {
     this.trackCliCommand({
       command: 'ai-gateway',
